# Improve performance for TimeSeriesResult.to_df